# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Flexibility:
    enum:
    - 0
    - 30
    - 60
    type: integer
    x-enum-varnames:
    - Exact
    - Approximate
    - Flexible
...
  domain.Ride:
    properties:
      Date:
        type: string
      DepartureAt:
        type: string
      Destination:
        type: string
      DriverName:
        type: string
      ExpiresAt:
        type: string
      FeedbackGiven:
        type: boolean
      Flexibility:
        $ref: '#/definitions/domain.Flexibility'
      ID:
        type: string
      Origin:
        type: string
      Phone:
        type: string
      PostedAt:
        type: string
    type: object
...
  handler.DeleteRideBody:
    properties:
      phone:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.InterestListItem:
    properties:
      id:
        type: string
      searcher_name:
        type: string
      searcher_phone:
        type: string
      status:
        type: string
    type: object
...
  handler.PostRideBody:
    properties:
      departure_at:
        type: string
      destination:
        type: string
      driver_name:
        type: string
      flexibility:
        type: integer
      origin:
        type: string
      phone:
        type: string
    type: object
  handler.PublicRequest:
    properties:
      DepartureAt:
        type: string
      Destination:
        type: string
      Flexibility:
        type: integer
      ID:
        type: string
      Origin:
        type: string
      SearcherName:
        type: string
    type: object
  handler.PublicRide:
    properties:
      Date:
        type: string
      DepartureAt:
        type: string
      Destination:
        type: string
      DriverName:
        type: string
      ExpiresAt:
        type: string
      FeedbackGiven:
        type: boolean
      Flexibility:
        type: integer
      ID:
        type: string
      InterestCount:
        type: integer
      Origin:
        type: string
      PostedAt:
        type: string
    type: object
  handler.SubscriptionBody:
    properties:
      auth:
        type: string
      endpoint:
        type: string
      p256dh:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
  description: Local ride-sharing notice board API.
  title: Go-Stop API
  version: "1.0"
paths:
...
  /rides:
    get:
      operationId: listRides
      parameters:
      - description: Origin filter
        in: query
        name: origin
        type: string
      - description: Destination filter
        in: query
        name: destination
        type: string
      - description: Departure time (RFC3339)
        in: query
        name: departure_at
        type: string
      - description: Search date (YYYY-MM-DD)
        in: query
        name: search_date
        type: string
      - description: Search time (HH:MM, local)
        in: query
        name: search_time
        type: string
      - description: Caller phone for my-rides mode
        in: header
        name: X-Phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.PublicRide'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - rides
    post:
      consumes:
      - application/json
      operationId: createRide
      parameters:
      - description: Ride to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PostRideBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Ride'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - rides
  /rides/{id}:
    delete:
      consumes:
      - application/json
      operationId: deleteRide
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Owner phone
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.DeleteRideBody'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - rides
    get:
      operationId: getRide
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Ride'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - rides
...
  /rides/{id}/interests:
    get:
      operationId: listRideInterests
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Ride owner phone
        in: header
        name: X-Phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.InterestListItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - interests
  /rides/{id}/requests:
    get:
      operationId: listRideRequests
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Ride owner phone
        in: header
        name: X-Phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.PublicRequest'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - rides
...
  /subscriptions:
    post:
      consumes:
      - application/json
      operationId: upsertSubscription
      parameters:
      - description: Push subscription
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SubscriptionBody'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - subscriptions
  /subscriptions/{phone}:
    delete:
      operationId: removeSubscription
      parameters:
      - description: Phone
        in: path
        name: phone
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - subscriptions
...
swagger: "2.0"